Add validation for SSP linked period of sickness
The need for this was identified during the fact-checking process.

The linked sickness period must be at least 4 calendar days long to count as a
Period of Incapacity for Work [1,2].

Although it's not very obvious from the code, this rule is already applied to
the *current* period of sickness. In the `must_be_sick_for_4_days?` question
the following two lines mean that the user is taken to the
`must_be_sick_for_4_days` outcome unless the current period of sickness is at
least 4 calendar days long:

        next_node_if(:has_linked_sickness?) { days_sick > 3 }
        next_node(:must_be_sick_for_4_days)

Although I think the use of `> 3` rather than `>= 4` and the use of the magic
number `3` are not very helpful in the context of the policy documentation, I've
decided to keep the new linked period of sickness code as similar as possible to
the old current period of sickness code.

Note that I elected to replace the `start_before_end` validation with the new
validation, because the former would never have been triggered once the latter
had been introduced.

Note also that I have confirmed with HMRC that this validation should not take
the employee's working pattern into account.

Fortuitously this new validation does not affect any of the combinations of
responses used in the regression tests.

[1]: https://www.gov.uk/guidance/statutory-sick-pay-manually-calculate-your-employees-payments#period-of-incapacity-for-work-piw
[2]: http://www.hmrc.gov.uk/manuals/spmmanual/spm10135.htm