Extract magic number into a named constant
Both the current period of sickness and the linked period of sickness are
Periods of Incapacity for Work (PIW) as defined in the policy [1,2]. Thus they
must both conform to the rule that a valid PIW must be at least 4 consecutive
calendar days long.

The change in the comparison from `> 3` to `>= 4` does not change the behaviour,
but means that the code more closely matches the policy documentation.

By extracting the magic number `4` into a single named constant in both places
we make it clearer what it means and that the two periods of sickness must
conform to the *same* rule.

[1]: https://www.gov.uk/guidance/statutory-sick-pay-manually-calculate-your-employees-payments#period-of-incapacity-for-work-piw
[2]: http://www.hmrc.gov.uk/manuals/spmmanual/spm10135.htm