Correctly pluralise values in calculate-your-holiday-entitlement
@erik-eide noticed that we were always displaying the plural of "shifts",
"days", "hours" and "minutes" in the outcomes of this Smart Answer.

This change uses the `pluralize` helper so that we correctly display the
singular or plural version of each of those words.

There were no changes to regression test artefacts, presumably because we only
ever test with quantities of more than 1.