Use next_node blocks in :purpose_of_visit?
Note the three sets of "rules" in the `next_node` block. I've had to split it
like this to replicate the existing behaviour. Splitting the rules like this has
meant that I've had to use `next` to return early from the `next_node` block in
the first `case` statement, and second `if` statement.