Use next_node block in :home_features_older?
I've replaced `define_predicate` with `next_node_calculation` in an attempt to
keep the code as similar as possible. I've had to update the names used by
`next_node_calculation` as the question marks cause problems when the state
object is converted to a hash in preparation for rendering the ERB templates.