Make SmartAnswer::I18nRenderer#value_for_interpolation private
I want to be able to change the `SmartAnswer::QuestionPresenter` to use the
`SmartAnswer::ErbRenderer` and so anything I can do to reduce the dependencies
on the `I18nRenderer` will make this easier.

This commit introduces a little duplication in the question presenter
sub-classes. I plan to fix that in a later commit.