Fix bug when dob_amount is more than 2 days in future
Partly fixes issue #2079.

This problem was introduced in f89112454eb899b8570c474f01db91cbd96862b2
(part of PR #2061) when we switched to using `next_node` with a block in
the `:dob_age?` question.

The implementation before f89112454eb899b8570c474f01db91cbd96862b2
relied on the `validate` block preventing the
`StatePensionAmountCalculator` being instantiated with a date of birth
that it couldn't handle (notably a date of birth more than two days in
the future). The switch to instantiating the calculator in the
`next_node_calculation` block meant that it was created before the
`validate` block is executed, which in turn meant that it could be
created with an invalid date of birth. The fix in this commit is to
instantiate the calculator in the `next_node` block so that we once
again rely on the `validate` block to reject invalid dates.

The test I've added is for the unchanged `validate` behaviour but it
gave me a failing test that I was able to use to drive the fix.

The problem ultimately stems from `StatePensionDateQuery#find_date`.
When the date is more than two days in the future, no data can be found
in either the static or dynamic pension data.

Note. The last entry in the
`StatePensionDateQuery#pension_dates_dynamic` array has an end date of
`Date.today + 1`. This explains why the error only occurs when the date
is more than two days in the future.

Note. I've had to change the `and` to `&&` when assigning a value to the
`near_pension_date` variable. Without this change, some of the
integration and regression tests for calculate-state-pension failed.