Add option to render questions using ERB template vs I18n
This makes it possible to switch between rendering questions using the
`SmartAnswer::I18nRenderer` or the `SmartAnswer::ErbRenderer` on a per question
basis.

Initially I tried to split the `SmartAnswer::QuestionPresenterTest` into two
contexts, but this didn't play well with the existing use of the `test` method
(i.e. not `should`), so I plumped for a new test case in a separate file.

Eventually we hope to do away with the `I18nRenderer` style and make the
`ErbRenderer` style the default. When this happens, we'll be able to replace
the `SmartAnswer::QuestionPresenterTest` with the new test case.

Note that I decided to use a symbol to lookup error message using ErbRenderer
to be consistent with the other calls to `#content_for`.