Treat error key consistently as a String
As far as I can tell, `SmartAnswer::State#error` is only ever set to be the
message from an `InvalidResponse` exception [1] i.e. it will always be a
`String`.

Also the key is always interpolated into the full i18n String key in the
renderer [2].

Thus it seems unnecessary and confusing to convert it to a symbol when calling
`QuestionPresenter#error_message_for` to lookup a specific error key, but not
when calling the same method to lookup the fallback error key, 'error_message'.

[1]: https://github.com/alphagov/smart-answers/blob/6302a552ab01e357513cc0b60f31cb08aa1d382e/lib/smart_answer/flow.rb#L125
[2]: https://github.com/alphagov/smart-answers/blob/7e26c554be106b2888f44bf92746e5fc7f6d819f/lib/smart_answer/i18n_renderer.rb#L22