Improve unit tests for QuestionPresenter#error
My main objective here was to cover the behaviour when an
`SmartAnswer::InvalidResponse` is raised with a text message and the default
error message in the i18n YAML file interpolates this error. Currently this
behaviour is only tested in the `InputValidationTest` integration test, but
there's so much going on in there it's hard to see the wood for the trees.

I'm not convinced that this behaviour is very sensible nor do I think it is used
anywhere in the published smart answer flows. However, it does appear as if
there are quite a few places where a custom String is set as the exception
message rather than an error message key. Since I believe the mechanism
described above is the only way such messages could currently be displayed to
the user, I'm loathe to remove it without providing a replacement.

I've also clarified that the final fallback error message is system-wide and not
just for the question's flow.