Remove deprecated Flow#section & #subsection_slug
Most of the code relating to these methods was removed in 2012 in the
following two commits:

* 8bee67ea43dc4de5e024342931c8028785d10f4d
* d397d13352b3832d263867ba742f256955272374

The deprecation warnings were left in place in case any un-merged flows used
either of the methods. I think it's safe to say that this danger has passed and
it's now safe to removed these methods.