Extract SmartAnswersControllerSampleWithMultipleChoiceQuestionFlow
Define an explicit flow class just for this test case to avoid sharing setup
with other tests i.e. make this test less brittle at the expense of more
verbosity.

Note that although it wasn't strictly necessary for this commit (because of the
`SmartAnswer::FlowRegistry.find`), I've changed the `id` parameter for all
requests to match the name of the new flow. In order to do this, I also had to
override `SmartAnswersControllerTestHelper#submit_response` and
`#submit_json_response`.