Avoid stubbing FlowRegistry in SmartAnswersControllerMultipleChoiceQuestionTest
It seems simpler and more realistic to use the real `SmartAnswer::FlowRegistry`
rather than stubbing everything. By using the methods in `FixtureFlowsHelper`
we can ensure that preloading of flows is disabled, thus only the flow under
test should be loaded.

Unfortunately we still need to setup the translation file
explicitly, because this is normally done globally and not by each flow.