Use FormattingHelper#format_money in SSP outcome
We're trying to standardize on using this method to format money objects.

The effect of this is to leave off the decimal point and the pence when the
value is a whole number of pounds.

@lutgendorff confirmed that this was probably just an accidental difference and
so I'm fixing it here.

Note that many of the regression test artefacts need to be updated, but I plan
to do that in a separate commit so it's easier to see what's going on.