ErbRenderer#content_for should always return HTML-safe string
The fact that all content has been rendered by an instance of `ActionView::Base`
means that it's safe to assume it's HTML-safe.

`GovspeakPresenter#html` was already being marking its output as HTML-safe, but
non-Govspeak content was not being marked as such.

It would probably make sense to move the call to `String#html_safe` in
`GovspeakPresenter#html` into `ErbRenderer#content_for`, but it's used in quite
a few places and I want to reduce the scope of this change, so this is a good
first step.