Extract ErbRenderer#single_line_of_content_for method
The main purpose of this is so that we can move the call to `String#chomp` into
the `ErbRenderer` in order to mark the result as HTML-safe.

In the long run, we could probably do without the call to `String#chomp`
altogether, but the trouble is that it would cause changes to most/all of the
regression test artefacts.