Add State#error and State#response
I'm planning to override `State#method_missing` so that it raises a
`NoMethodError` if you attempt to read an attribute value without first setting
it. We always want to be able to read `error` and `response` irrespective of
whether they've been set.

I initially tried using `attr_reader` but that means that `error` and `state`
are left out of the hash returned from `State#to_hash`. Setting them in this
call to `super` ensures that `#to_hash` works as expected.