Update regression test artefacts
The page `title` and `h1` elements on all question pages is obtained from the
start node title. This is now *correctly* being marked as HTML-safe when it
comes back from the `ErbRenderer` and so the apostrophes are no longer being
HTML-escaped.

Note that @chrisroos & I were confused for a while because the old HTML-escaped
behaviour was not visible in development or production. We're pretty sure this
is because Slimmer uses Nokogiri to extract a fragment of HTML from the page
and then re-inserts it in a new "layout". We suspect this process does some kind
of normalisation which means the escaped apostrophes are unescaped.