Convert flow-sample to use question ERB templates
I ran the following command:

    $ rails r script/convert-question-i18-yaml-to-erb-templates.rb test/fixtures/flow-sample

This flow is used in two unit tests:

* `SmartAnswer::FlowRegistryTest`

This test didn't need any changes, because it wasn't making use of any of the
I18n content.

* `FlowRegistrationPresenterTest`

Even though this unit test, I elected to use the `FixtureFlowsHelper` methods in
order to setup the template directory for the question nodes. It would probably
be better to make it possible to configure the template directory when
instantiating a flow, but that's a job for another day.

Now that ERB tags can be used for interpolation, it's possible to have a chain
of method calls instead of a single level which was all that was possible with
I18n interpolation. I've therefore changed the flow fixture to add test
coverage for this scenario and re-used the same technique as for the
`GraphPresenter` to handle this scenario i.e. using `MethodMissingHelper` and
`MethodMissingObject`. Since we're generating content that is intended to be
used in a search index, in this case we don't want to display the missing
variable names; we just need to avoid raising an exception.

To avoid having to create a whole new flow for the one interpolation test,
I opted to add a question to the flow within that one test.

I've renamed the tests which talk about translations since we're now using ERB
rendering not I18n translation.