Convert graph flow to use question ERB templates
I ran the following command:

    $ rails r script/convert-question-i18-yaml-to-erb-templates.rb test/fixtures/smart-answers-controller-sample

Even though this flow is used in a unit test, I elected to use the
`FixtureFlowsHelper` methods in order to setup the template directory for the
question nodes. It would probably be better to make it possible to configure
the template directory when instantiating a flow, but that's a job for another
day.

I've implemented a solution for coping with interpolating values which are not
available from the state object. The `MethodMissingHelper` is mixed into the
view context and builds an instance of `MethodMissingObject`. The latter
implements a `#to_s` method which returns the name of the method.

Now that ERB tags can be used for interpolation, it's theoretically possible to
have a chain of method calls instead of a single level which was all that was
possible with I18n interpolation. The above solution only deals with a single
level.

I've also removed all the i18n-related code from the test.