Handle arbitrary levels of interpolation in GraphPresenter#labels
Now that ERB tags can be used for interpolation, it's possible to have a chain
of method calls instead of a single level which was all that was possible with
I18n interpolation. I've therefore changed the flow fixture to add test
coverage for this scenario and enhanced `MethodMissingObject` so that it can
handle that scenario.