Remove brittle assertion checking non-deterministic behaviour
This assertion doesn't seem to be buying us much and has recently started to
cause intermittent CI build failures like this one [1].

I think the problem is that this test was relying on all fixture flows except
`flow-sample` to be marked as `draft` and/or that the enumeration of files in
the `load_path` directory is deterministic across different filesystems.

The fixture flows recently added for the various `SmartAnswersController`
functional tests are *not* marked as `draft` and I think this is why the builds
have started to fail occasionally.

Since the assertion isn't really buying us much, @chrisroos & I have agreed that
it should be removed.

[1]: https://ci.dev.publishing.service.gov.uk/job/govuk_smartanswers_branches/3376/console