Update regression test artefacts
The change to ERB templates has introduced a few extra newlines in the
question page HTML, but none of them affect the content viewable by the user.

I used the following command to hide these changes:

    $ git diff --ignore-blank-lines

The change to ERB templates has also removed the trailing newline for question
attributes which were using a `|`-prefixed multi-line YAML string. Even though
the newline has been reproduced faithfully in the corresponding ERB template, it
has been lost by one/both of the calls to `ErbRenderer#normalize_blank_lines`
and/or `String#chomp`.

I feel as if the resulting HTML is actually cleaner and so I'm happy to update
the regression test artefacts accordingly.