Use ERB tags to interpolate values in template
These were not dealt with automatically by the conversion script.

I've also elected to inline the phrases from the i18n YAML file into the flow
`precalculate` blocks. Note that I've had to mark on of them as HTML-safe to
avoid an apostrophe being HTML-escaped. This isn't ideal, but I'd hope that
these strings would move into the ERB template or a helper method where this
would be easier to deal with.