Fix GraphPresenter to work with both i18n & ERB
In #2103 I forgot that we still need to support flows/questions with their
content in i18n YAML files vs ERB templates.

In this commit I've renamed the `GraphPresenterTest` to
`GraphPresenterWithErbRendererTest` and reinstated the earlier
version of `GraphPresenterTest` to give suitable test coverage.

I've also reinstated the i18n YAML file and used an explicit call to
`Flow#use_erb_templates_for_questions` in `GraphPresenterWithErbRendererTest` to
switch over to ERB rendering.

I've avoided needing to handle `I18n::MissingInterpolationArgument` by passing
in an empty `Hash` as the `state` when instantiating the `QuestionPresenter`.