Fix duplicate test context
This was causing me problems when I tried to make the change to use an ERB
template for the question, because some of the setup in the second context was
needed by the tests in the first context. This is because when any question
attribute is accessed, the whole ERB template has to be rendered.