Don't use multi-line YAML strings for single-line content
This results in more aesthetically pleasing HTML and should make it easier to
convert the flow to use ERB templates for questions.

There is one `hint` in the YAML file which is two sentences on two different
lines with a blank line in between. Presumably the idea was that this would be
converted into two paragraph tags by Govspeak. However, attributes like `hint`
have never (?) been processed by Govspeak, so this would always have been
rendered as a single paragraph tag. Thus it's reasonable for me to combine it
all onto a single line.

I've updated the regression test artefacts and checksums accordingly.