Move data_partials/_button.erb
The smart_answer_flows/data_partials directory is a hangover from before the
introduction of ERB templates. The smart_answer_flows/shared directory is where
shared partials now live.

This partial is only used by the register-a-birth and register-a-death Smart
Answers so I've placed it in the shared/births_and_deaths_registration
directory.

I've added the .govspeak extension so that it's name is consistent with the
other partials in this shared directory.