Move data_partials/_overseas_passports_embassies.erb
The smart_answer_flows/data_partials directory is a hangover from before the
introduction of ERB templates. The smart_answer_flows/shared directory is where
shared partials now live.

This partial is used by 5 Smart Answers so I've placed it in the shared/
directory rather than a subdirectory of shared/.

I've added the .govspeak extension so that it's name is consistent with the
other partials in this shared directory.