Avoid using next_node_if from next_node
The use of predicates to define next node rules was deprecated a while back and
we're now in the process of removing this predicate code.

I've retained the ability to call `next_node :next_node_name` by setting
`@default_next_node_function` to a `proc` that returns the `:next_node_name`.
This replicates the behaviour of calling `next_node` with a block directly.