Avoid using use_erb_templates_for_questions in test flows
It's now the default to use ERB templates for questions so we no longer
need to specify it explicitly.

I've reversed the assertions in the "change default renderer for
questions" test in test/unit/flow_test.rb. It now asserts that the first
question uses ERB and that the second has been switched to use I18n.

I've switched to using `Flow.build` instead of `Flow.new` and
`Flow#define` in flow_registration_presenter_with_erb_renderer_test and
graph_presenter_with_erb_renderer_test. This is because it's no longer
necessary to call `#use_erb_templates_for_questions` between
instantiation and the call to `#define`.