Add unit tests for FlowPresenter#presenter_for
It's pretty surprising that such a critical class (`FlowPresenter`) had no unit
tests prior to this.

I want to make changes to the `#presenter_for` method, so I've only added tests
for this method for now in order to give myself confidence that my upcoming
changes do not break any existing behaviour.

Unfortunately I had to add a single question to the flow, because
`Flow#start_state` (which is used in calculating `FlowPresenter#current_state`)
relies on there being at least one question in the flow.

Note that these tests do not endorse the existing behaviour.