Add unit tests for FlowPresenter#presenter_for
It's pretty surprising that such a critical class (`FlowPresenter`) had no unit
tests prior to this.

I want to make changes to the `#presenter_for` method, so I've only added tests
for this method for now in order to give myself confidence that my upcoming
changes do not break any existing behaviour.

Note that these tests do not endorse the existing behaviour.