Memo-ize FlowPresenter#presenter_for
`FlowPresenter#presenter_for` is called many times on the same instance within
a single request; most notably from within `FlowPresenter#collapsed_questions`.
This meant that we were unnecessarily creating many instances of the subclasses
of `NodePresenter`.

An instance of `SmartAnswer::ErbRenderer` is created for each of these and an
instance of `ActionView::Base` is created for each of those.

We can reuse the relevant node presenter by memo-izing them in a `Hash` keyed on
the node's name. It's safe to use the node's name as the key, because node names
are unique for a flow i.e. it's not possible to have different node types with
the same name within the same flow (see `Flow#add_node`).

There's no need to key on the value returned by `FlowPresenter#current_state` as
well, because this is itself memo-ized and is therefore always the same for a
given instance of `FlowPresenter`.

It's safe to memo-ize the node presenters, because a new `FlowPresenter`
instance is created for every request within the
`SmartAnswersController#find_smart_answer` method.