Memo-ize FlowPresenter#start_node
`FlowPresenter#start_node` is called several times on the same instance within
a single request. This meant that we were unnecessarily creating several
instances of `StartNodePresenter`.

An instance of `SmartAnswer::ErbRenderer` is created for each of these and an
instance of `ActionView::Base` is created for each of those.

We can reuse the start node presenter by memo-izing it. It's safe to do this,
because a new `FlowPresenter` instance is created in every request within the
`SmartAnswersController#find_smart_answer` method.