Use current_node vs questions.first in content partial
Note that more generally `SmartAnswerPresenter#current_node` can be a
`StartNodePresenter` or an `OutcomePresenter`, but in this case we're "inside"
the `else` block for `if ! @presenter.started?` and the `else` block of
`if @presenter.finished?` and so it will always be a `QuestionPresenter` or
one of its subclasses.

It's now safe to remove the redundant `SmartAnswerPresenter#questions` method.