Stop passing number into question partial templates
These partials are only ever rendered with `number` set to `nil`, so we can
simplify the code by not passing it in and assigning `number = nil` inside
the partial templates. Note that not all question partial templates use this
`number` local variable, so I haven't bothered assigning it in those.