Update regression test artefacts for question pages
Many of the commits in this pull request change the whitespace in the
regression test artefacts for question pages, but the idea is that none of them
change the structure of the HTML.

In this commit, I've re-generated the the regression test artefacts for all the
question pages. I used two methods to convince myself that all the changes are
whitespace changes.

* Method 1
  * Run the following command:

    git diff --ignore-blank-lines --ignore-all-space

* Method 2
  * Install HTML Tidy [1]
  * Create a new git worktree for the branch
  * Update the regression test artefacts for all the question pages
  * Run the following command in both worktrees (master & branch):

    find test/artefacts -name "*.html" -exec tidy -modify --sort-attributes alpha --wrap 0 {} \;

  * Run the following command in the master worktree:

    diff -r test/artefacts ../<path-to-worktree-for-branch/test/artefacts

Both methods show no differences and so I'm happy that these changes to the
regression test artefacts are all only insignificant whitespace changes.

[1]: http://www.html-tidy.org/