Remove conditional selection of gem versions/paths
Developers can easily modify the Gemfile to include a path: <path> or git: <uri> to particular version of a gem if they want to work towards that rather than setting ENV variables which are very easy to forget to unset.

Also, allowing ENV variables to be set via a script or puppet, to change the version of a gem used in a hosted environment such as preview or production, where developers can’t see the values, is potentially a little confusing when trying to provide support.