Extend landlord immigration check calculator
Add valid countries to the landlord immigration check calculator to extend the area to include all of England now that the pilot is almost over.

Although the countries array only has one element at this stage, reusing the same structure as when checking if a postcode is in a valid Borough hopefully makes hopefully makes the code a bit easier to read and extend.