Update outcome for register a birth in Cambodia and PNG
Remove the exception for Cambodia and Papua New Guinea. This is the first paragraph and currently reads:

'Registration can take up to 3 months - your application will be sent to the local British embassy, consulate or high commission where your child was born to be registered there.'

Rest of text remains so both become same as all other ORUs.