Add postcode question to benefit cap calculator
With the introduction of London and non-London benefit cap rates, this adds the rates in a data file and a postcode question to the flow, so we can query if what rate to apply. With the possibility of further devolution of rates across the UK, the data file can be extended to include other areas.