Update calculator to allow optional param to be nil
The state_pension_age_calculator can be used for calculating the stage_pension_age, and the bus pass age. As the bus pass age defaults to :female for gender, this gender param to the new method can be optional.

With gender an optional param, changing gender from a read to a read/write accessor allows its’ value to be set in subsequent questions.