Update calculator for new outcomes in State Pension Age
Removes unused methods in the state pension age calculator which were a carry over from calculating the state pension amount, and introduces new methods for showing the state pension age and bus pass age outcomes.

With the StatePensionDateQuery class now having test coverage for the pension age data, it removes the tests duplicating testing this data and adds tests around the functionality of the calculator, including edge conditions that flip boolean results of a method call.