Add empty MarriageAbroadCalculator
Instantiate it in a `next_node_calculation` block in the first question
so that it's immediately available to use.

I've added an empty test placeholder ready for when I add behaviour to
the MarriageAbroadCalculator.

I've added the marriage_abroad_flow_test so that I can assert that we're
instantiating the calculator in the first question. I had to stub
`WorldLocation.all` and `WorldLocation.find` although I have a feeling
we might be doing this slightly differently elsewhere.