Inline :ss_affirmation
This `next_node_calculation` was only being used in a single place and I
don't think it was adding much to the readability of the code.

In addition, the conditionals in the `next_node` block were using a mix
of these `next_node_calculations` and accessing the `data_query`
directly. I think having a single way of doing things is beneficial even
if things become slightly less readable after removing the
`next_node_calculation`s.