Inline :consular_cni_residing_in_third_country
I wasn't sure whether to inline this method or not, as it's possible
we're losing something by removing the
`:consular_cni_residing_in_third_country` label. Having said that, I'm
not entirely sure that the label has been updated as the conditional has
been added to (e.g. with the test for Kosovo hosting the ceremony).

In the end I decided to inline it as this was the final
`next_node_calculation` in this question. The conditionals in the
`next_node` block were using a mix of these `next_node_calculations` and
accessing the `data_query` directly. I think having a single way of
doing things is beneficial even if things become slightly less readable
after removing the `next_node_calculation`s.