Inline :consular_cni_residing_in_third_country
I wasn't sure whether to inline this method or not, as it's possible
we're losing something by removing the
`:consular_cni_residing_in_third_country` label. Having said that, I'm
not entirely sure that the label has been updated as the conditional has
been added to (e.g. with the test for Kosovo hosting the ceremony).

In the end I decided to inline it as this was the final
`next_node_calculation` in this question.