WIP: Fix bug in validation of linked PIW end date
The manual calculation explanation says that the gap between the current PIW and
the linked PIW must be 8 weeks (56 days) or less. I think this means that the
earliest end date for the linked PIW is 8 weeks and 1 day earlier than the
current PIW start date, but I need to confirm this with the Content Team and/or
the relevant government department.