Fix bug in validation of linked PIW end date
The manual calculation explanation [1] says:

> Link PIWs and treat as 1 PIW if the gap between them is 8 weeks (56 days) or
> less.

This means that there can be up to 56 days in *between* the two PIWs i.e. not
including the last day of the previous PIW nor the first day of the current PIW.

In the integration test the current PIW starts on 21/05/2015 and consulting the
HMRC table [2], the previous PIW is linked if it starts on or after 25/03/2015.

Previously the validation was incorrectly including both the last day of the
previous PIW and the first day of the current PIW in the gap. In order not to
include these, the simplest way is to subtract 8 weeks and 1 day (57 days) from
the date of the start of the current PIW. If the previous PIW starts on or after
this date then it is considered as linked.

This commit corrects the integration test and fixes the implementation.

I did also contemplate changing the names of the `must_be_within_eight_weeks`
validation block and the `within_eight_weeks_of_current_sickness_period?`
method. However, I decided not to do so for now, because they aren't very
misleading, just a slightly imprecise. And in any case I hope to incorporate the
work I did in #2280 to improve matters in this area.

[1]: https://www.gov.uk/guidance/statutory-sick-pay-manually-calculate-your-employees-payments#linking-piws
[2]: https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/406561/April_2015_SSP_PIW_tables.pdf