Add MarriageAbroadCalculator#world_location
I've moved the behaviour from the `location` calculate block to this new
method.

I don't think the `#world_location` method should be raising an
`InvalidResponse` exception if `WorldLocation#find` returns `nil`. I
think it should probably raise a custom error (e.g. "Data not found for
the chosen ceremony country" that's then handled somehow in the front
end). Note that the `country_select` question itself already handles the
case where the ceremony country doesn't exist in the list of countries
(this can happen if the URL is manually updated).

I'm not memoising the `world_location` because `WorldLocation` already
includes a cache to minimise the number of network requests.