Add MarriageAbroadCalculator#fco_organisation
I tried to write the tests for this new method using the
`worldwide_api_xxx` helper methods. I struggled to come up with a good
way to test that `#fco_organisation` was returning the organisation I
was interested in, which is why I ended up stubbing
`WorldLocation.find`. If this feels like a good approach then I'll
probably update the tests for `#world_location` so that they also use
this sort of stubbing instead of the `worldwide_api_xxx` helper methods.