Add MarriageAbroadCalculator#marriage_and_partnership_phrases
I've moved the behaviour from the `marriage_and_partnership_phrases`
calculate block to this new method.

I've replicated this method as-is at the moment although I can imagine
it changing/going away in future as its purpose becomes clearer.

I don't fully understand the methods on the MarriageAbroadDataQuery so I
can't be sure that my test names are accurate.

I've used stub.quacks_like(<instance-of-object>) in some of these tests.
This is instead of using a real instance of the object, or of using a
plain stub. It gives me confidence that I'm not accidentally relying on
behaviour in the real object, and that I'm not stubbing/calling a method
that doesn't exist on the real object.