Add MarriageAbroadCalculator#marriage_and_partnership_phrases
I've moved the behaviour from the `marriage_and_partnership_phrases`
calculate block to this new method.

I've replicated this method as-is at the moment although I can imagine
it changing/going away in future as its purpose becomes clearer.

I don't fully understand the methods on the MarriageAbroadDataQuery so I
can't be sure that my test names are accurate.