Add new outcome_work_waiver to check-uk-visa flow
All visitors in countries participating in the electronic visa waiver
program, coming to the UK for work for less than six months should
see a work specific outcome related to the EVW program.

Add integration tests covering the flow to the new outcome_work_waiver
outcome in the check-uk-visa flow for a country participating in the
EVW program.

Add the new outcome_work_waiver to the expected results for an Oman
visitor coming to the UK to work less than 6 months.