Setup abc tests for 3 smart-answers
We want to run abc tests for three smart-answers on the text in the
"Start Now" button.  We're using both content experiments (configured the
contentExperimentId for the test and variantIds for each cohort) and a
custom dimension to track the data as this gives us total flexibility
for tracking things.