Add new outcome_study_waiver to check-uk-visa flow
Visitors from countries participating in the electronic visa waiver
program, coming to the UK to study for less than six months should see
a study specific outcome related to the EVW program.

Add context to assert user is shown the the new outcome node
outcome_study_waiver when coming to the UK to study for less than 6 months.

Remove existing test `test venezuela special outcome - study -
less than six months`, as this is now captured in the new tests
grouped together in a context related to EVW countries, and also because
it’s incorrectly named, with Venezuela not being a country participating
in the EVW program.

Update the expected outcome for Oman, study, less than 6 months in
the regression tests to the new outcome_study_waiver outcome.