Add integration test to check-uk-visa for new outcome
Add context to assert user is shown the the new outcome node 
outcome_study_waiver when coming to the UK to study for less than 6 months.

Remove existing test `test venezuela special outcome - study -
less than six months`, as this is now captured in the new tests
grouped together in a context related to EVW countries, and also because 
it’s incorrectly named, with Venezuela not being a country participating
in the EVW program.