Add new outcome_school_waiver to check-uk-visa flow
Visitors from countries participating in the electronic visa waiver
program, coming to the UK to visit a child in school should see
a visiting a child in school specific outcome related to the EVW
program.

The commit also moves all `calculator.school_visit?` logic into one
branch of the `next_node` block of `purpose_of_visit?` question.

As this question is asking the purpose of the visit, I think the first
level of the of the `if` statement should be evaluating this value, and
sub-level checking for special conditions applying to the purpose of
visit inside these as nested if/else statements to lead to alternative outcomes.

Add test to assert a visitor from a EVW program member country sees
a `outcome_school_waiver` outcome when visiting a child in school.

Remove Oman test leading to the DATV specific outcome for visiting
a child in school which no longer applies to visitors from EVW
countries.

Update the expected outcome for Oman, visit child in school in regression
to to the new `outcome_school_waiver` outcome.